pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "open-secretary"
version = "0.1.0"
description = "Daily calendar summary to Telegram using Gemini"
requires-python = ">=3.11"
dependencies = [
  "APScheduler>=3.10.4,<4.0",
  "icalendar>=5.0.12",
  "requests>=2.31.0",
  "google-genai>=0.5.0",
  "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"
